offerBeam changes, network manager has been changed

+++ BizrateRewards/Classes/MVC/Controllers/SurveyViewController/FinishSurveyController/BZRFinishSurveyController.m
@@ -65,7 +65,7 @@
 
 - (void)setupObtainedPointsText
 {
-    self.obtainedPointsLabel.text = [NSString stringWithFormat:@"%i %@", self.passedSurvey.surveyPoints, self.obtainedPointsText];
+    self.obtainedPointsLabel.text = [NSString stringWithFormat:@"%li %@", (long)self.passedSurvey.surveyPoints, self.obtainedPointsText];
 }
 
 @end
